refactor: move construct sql query function into helpers library
tidy ups
